requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==1.10.13
requests==2.31.0
beautifulsoup4==4.12.2
python-multipart==0.0.9
google-generativeai==0.4.1
